Add icons to site legend
